Explicitly specify direct dependencies
Instead of relying on `swagger-ui` including its dependencies (with
global exports) and relying on their existence in 'custom' code, this
commit alters `package.json` to explicitly list the (direct)
dependencies being used, and pinning these to the version as included in
the `swagger-ui` distribution (except for JQuery: `jquery-1.8.0` is no
longer available on NPM, so I included the closest available version
instead).

Next, all these modules are aliased in the Webpack configuration, so
that instead of the libraries retrieved from NPM are not used, but the
files as embedded in the `swagger-ui` distribution are used instead.

Finally, except for the `swagger-ui-lib` module, all former `require`s
are turned into ES6-style 'global' imports (due to the use of
`script-loader` for the dependent libraries, named imports are not
possible).

Short things short: dependencies are now explicit, but the bundled
versions are still used.