Build `swagger-ui` 'from source'
Instead of packaging the 'built' sources from `swagger-ui`, include it
as a 'source' package.

As a drawback, because there's no single entry-point, this requires
listing all source files separately and a little work-around for
`SwaggerClient` to be defined.

A positive side-effect is we no longer need to use the 'script-loader'
for the `swagger-ui` modules (but still need it for the dependencies,
for now).