Library'fy the package
Instead of shipping this package as a bundle including all of SwaggerUI,
this commit turns the package into a library, consisting of two parts:

- `SwaggerAws.Authorization`, which contains a
  `swagger-client`-compatible authorization provider for AWS4 signing of
  requests.
- `SwaggerAws.Ui`, which contains all code required to add AWS4
  authorization support to SwaggerUI using monkey-patching. This depends
  on `SwaggerAws.Authorization`.

To add this to a SwaggerUI installation, the following changes should be
made:

- Add a `script` tag loading `SwaggerAws.Authorization` to the page.
- Add a `script` tag loading `SwaggerAws.Ui` to the page. Note: this tag
  should come *after* the loading SwaggerUI and all of its dependencies.
- Call `SwaggerAws.Ui.patch()` to monkey-patch AWS4 signing support into
  SwaggerUI.

The libraries coming with SwaggerUI and also used by this package are
listed as `peerDependencies`. Sadly enough, the exact version of JQuery
coming with SwaggerUI is not available on NPM so one is stuck with a
warning about the correct version not being installed. This shouldn't
cause any issues in a browser context though.

The version of this package tracks the version of SwaggerUI it's
compatible/tested with.