Replace `crypto` with `crypto-js`
The `aws4` package uses the `crypto` package (which in a browser becomes
`crypto-browserify`) for SHA256 and HMAC calculations. Sadly enough, the
API of this package requires it to be fairly monolithic, such that
tree-shaking is unable to only include the code (algorithms) that's
actually used.

As a result, we end up with a bundle of +- 340kB in size, containing all
kinds of crypto primitives, hash functions, elliptic curve functions,
ciphers, etc.

This commit creates a module which exposes the same API as `crypto`
(insofar the API is used by `aws4`), but using `crypto-js` which allows
for more fine-grained inclusion of desired functionality.

Finally the `crypto` module is replaced by the mock module using
Webpack's `alias` feature.

This brings down script size from +- 340kB to +- 54kB.