Test: Don't include `host` in API spec
Removing this simplifies things *a lot*: according to the Swagger spec,
if a value for `host` is not provided, the host serving the document
shall be used. This is exactly what we need for various testing
scenarios: no more need to retrieve, pass through and patch on addresses
and ports!