Test: Add headless mode for e2e tests
This mode requires no browsers to be installed on the host system,
including no display server. This allows to run the e2e tests in a CI
environment.

This feature *does* rely on Docker and `docker-compose` to be available.

Note: currently the e2e scenarios are only ran on Google Chrome. For
some reason, when using this approach, the tests fail on Firefox (the
login dialog doesn't close, key input fields are in error state.
Unclear for now). The `e2e` NPM target works without problem though...